fix: improve bannerType logic and share it between banners
