+++ ts/features/itwallet/common/components/discoveryBanner/ItwDiscoveryBannerOnboarding.tsx
@@ -2,8 +2,6 @@ import { Banner } from "@pagopa/io-app-design-system";
 import { useRoute } from "@react-navigation/native";
 import { memo, useCallback, useMemo } from "react";
 import { StyleSheet, View } from "react-native";
-import { pipe } from "fp-ts/function";
-import * as O from "fp-ts/Option";
 import I18n from "../../../../../i18n";
 import { useIONavigation } from "../../../../../navigation/params/AppParamsList";
 import { useIOSelector } from "../../../../../store/hooks";
@@ -14,7 +12,7 @@ import {
 } from "../../../analytics";
 import { ITW_ROUTES } from "../../../navigation/routes";
 import { isItwDiscoveryBannerRenderableSelector } from "../../store/selectors";
-import { itwIsWalletInstanceRemotelyActiveSelector } from "../../store/selectors/preferences.ts";
+import { useItwDiscoveryBannerType } from "../../hooks/useItwDiscoveryBannerType.ts";
 
 /**
  * ITW discovery banner to be displayed in the wallet card onboarding screen
@@ -22,13 +20,9 @@ import { itwIsWalletInstanceRemotelyActiveSelector } from "../../store/selectors
 const ItwDiscoveryBannerOnboarding = () => {
   const navigation = useIONavigation();
   const route = useRoute();
-
+  const bannerType = useItwDiscoveryBannerType();
   const isBannerRenderable = useIOSelector(
     isItwDiscoveryBannerRenderableSelector
-  );
-
-  const isWalletRemotelyActive = useIOSelector(
-    itwIsWalletInstanceRemotelyActiveSelector
   );
 
   const trackBannerProperties = useMemo(
@@ -55,14 +49,6 @@ const ItwDiscoveryBannerOnboarding = () => {
     });
   };
 
-  const shouldRender = pipe(
-    O.fromNullable(isWalletRemotelyActive),
-    O.fold(
-      () => false,
-      () => isBannerRenderable
-    )
-  );
-
   const bannerConfig = {
     onboarding: {
       content: I18n.t("features.itWallet.discovery.banner.home.content")
@@ -74,12 +60,11 @@ const ItwDiscoveryBannerOnboarding = () => {
     }
   };
 
-  const bannerType = isWalletRemotelyActive ? "reactivating" : "onboarding";
-  const { content } = bannerConfig[bannerType];
-
-  if (!shouldRender) {
+  if (!isBannerRenderable || !bannerType) {
     return null;
   }
+
+  const { content } = bannerConfig[bannerType];
 
   return (
     <View style={styles.wrapper}>
